do only exists

+++ server.py
@@ -24,12 +24,26 @@ class ValidateError(Exception):
 def handler(job):
     try:
         print("started")
-        with open("runpod.yaml", "r") as f:
-            config = yaml.safe_load(f)
-            config['mixed_precision'] = job['input']['train']['mixed_precision']
-            os.makedirs("/root/.cache/huggingface/accelerate/", exist_ok=True)
-            with open("/root/.cache/huggingface/accelerate/default_config.yaml", "w") as fw:
-                yaml.dump(config, fw)
+        if "input" not in job:
+            raise ValidateError("input is required")
+        if "type" not in job["input"]:
+            raise ValidateError("type is required")
+        if "train" not in job["input"]:
+            raise ValidateError("train is required")
+        if os.path.exists("runpod.yaml"):
+            with open("runpod.yaml", "r") as f:
+                config = yaml.safe_load(f)
+                config['mixed_precision'] = job['input']['train']['mixed_precision']
+                os.makedirs("/root/.cache/huggingface/accelerate/", exist_ok=True)
+                is_existing = os.path.exists("/root/.cache/huggingface/accelerate/default_config.yaml")
+                if not is_existing:
+                    with open("/root/.cache/huggingface/accelerate/default_config.yaml", "w") as fw:
+                        yaml.dump(config, fw)
+                else:
+                    print("default_config.yaml already exists")
+        else:
+            print("runpod.yaml not found")
+        
         job_input = job["input"] # Access the input from the request.
         working_repo = job_input["working_repo"] if "working_repo" in job_input else ""
         repo_dir = job_input['output']['model_name']
